+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPNFMM89FJPKYP0953VV7QBK"
+        "01KPNJFE3WWXWNE3DKZARC3XDJ"
     )
     r"""The unique identifier of the data source"""