+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPNFMM88HP0JTSBEZTTH1KWB"
+        "01KPNJFE3V3AVJ9HQCC93EQH6N"
     )
     r"""The unique identifier of the data source"""