+++ packages/orq-rc/src/orq_ai_sdk/models/retrieveidentityop.py
@@ -189,7 +189,7 @@ class RetrieveIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-04-20T13:00:08.690Z")
+    updated: Optional[datetime] = parse_datetime("2026-04-20T13:49:43.154Z")
     r"""The date and time the resource was last updated"""
 
     metrics: OptionalNullable[RetrieveIdentityMetrics] = UNSET